changes in Simple Parser for Text File Lines - changed name of parser to LineBasedParser - removed static
